# [openrr-planner] Parallelize get_reachable_region