Before:
```
start reachable region calculation
[openrr-planner/src/ik.rs:128] now.elapsed() = 98.987975589s
finished reachable region calculation
```

After(this PR):
```
start reachable region calculation
[openrr-planner/src/ik.rs:137] now.elapsed() = 18.497009864s
finished reachable region calculation
```

FYI(this PR + [patch to keep order](https://github.com/openrr/openrr/commit/b69faa5f88a7e2a5f19f87a5cdd33324d31bfe0d)):
```
start reachable region calculation
[openrr-planner/src/ik.rs:142] now.elapsed() = 19.616745919s
finished reachable region calculation
```

FYI(this PR + https://github.com/openrr/k/pull/68):
```
start reachable region calculation
[openrr-planner/src/ik.rs:137] now.elapsed() = 14.921759368s
finished reachable region calculation
```

(Benchmarked on MacBook Pro (Intel Core i7-9750H))